Creates PatchingUtilities to help with patching methods (#8)
* Creates PatchingUtilities, including transpiler used in AlphaAnimals and RNG patcher used in AlphaAnimals and VFEF

* Updated VFEPower to use PatchingUtilities as well
